Test: add a trace test.
As Maven build process is not as customizable as a makefile, we will
probably have to remove automatic traces activation in the process.
A new test for traces mechanism ensures that the code is covered by
tests.

Testing Done: Build, read the coverage report, see that TraceTest is
fully covered and the Trace class too.
Issue Number: 6